bot/requirements.txt
aiopg==1.3.4
discord.py==1.7.3
peewee==3.15.1
peewee-async==0.8.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
typing_extensions==4.3.0
